+++ crates/nerd-font-symbols/src/dev.rs
@@ -1,131 +1,301 @@
+/// ``
+pub const DEV_AARCH64: &str = "";
+/// ``
+pub const DEV_ADONISJS: &str = "";
+/// ``
+pub const DEV_AFTEREFFECTS: &str = "";
+/// ``
+pub const DEV_AKKA: &str = "";
+/// ``
+pub const DEV_ALGOLIA: &str = "";
+/// ``
+pub const DEV_ALPINEJS: &str = "";
+/// ``
+pub const DEV_AMAZONWEBSERVICES: &str = "";
+/// ``
+pub const DEV_ANACONDA: &str = "";
 /// ``
 pub const DEV_ANDROID: &str = "";
+/// ``
+pub const DEV_ANDROIDSTUDIO: &str = "";
 /// ``
 pub const DEV_ANGULAR: &str = "";
+/// ``
+pub const DEV_ANGULARJS: &str = "";
+/// ``
+pub const DEV_ANGULARMATERIAL: &str = "";
+/// ``
+pub const DEV_ANSIBLE: &str = "";
+/// ``
+pub const DEV_ANTDESIGN: &str = "";
+/// ``
+pub const DEV_APACHE: &str = "";
+/// ``
+pub const DEV_APACHEAIRFLOW: &str = "";
+/// ``
+pub const DEV_APACHEKAFKA: &str = "";
+/// ``
+pub const DEV_APACHESPARK: &str = "";
+/// ``
+pub const DEV_APL: &str = "";
 /// ``
 pub const DEV_APPCELERATOR: &str = "";
 /// ``
 pub const DEV_APPLE: &str = "";
-/// ``
-pub const DEV_APPSTORE: &str = "";
-/// ``
-pub const DEV_APTANA: &str = "";
-/// ``
-pub const DEV_ASTERISK: &str = "";
-/// ``
-pub const DEV_ATLASSIAN: &str = "";
+/// ``
+pub const DEV_APPWRITE: &str = "";
+/// ``
+pub const DEV_ARCHLINUX: &str = "";
+/// ``
+pub const DEV_ARDUINO: &str = "";
+/// ``
+pub const DEV_ARGOCD: &str = "";
+/// ``
+pub const DEV_ASTRO: &str = "";
 /// ``
 pub const DEV_ATOM: &str = "";
-/// ``
-pub const DEV_AWS: &str = "";
+/// ``
+pub const DEV_AWK: &str = "";
+/// ``
+pub const DEV_AXIOS: &str = "";
+/// ``
+pub const DEV_AZURE: &str = "";
+/// ``
+pub const DEV_AZUREDEVOPS: &str = "";
+/// ``
+pub const DEV_AZURESQLDATABASE: &str = "";
+/// ``
+pub const DEV_BABEL: &str = "";
 /// ``
-pub const DEV_BACKBONE: &str = "";
-/// ``
-pub const DEV_BING_SMALL: &str = "";
-/// ``
-pub const DEV_BINTRAY: &str = "";
+pub const DEV_BACKBONEJS: &str = "";
+/// ``
+pub const DEV_BALLERINA: &str = "";
+/// ``
+pub const DEV_BAMBOO: &str = "";
+/// ``
+pub const DEV_BASH: &str = "";
+/// ``
+pub const DEV_BEATS: &str = "";
+/// ``
+pub const DEV_BEHANCE: &str = "";
 /// ``
 pub const DEV_BITBUCKET: &str = "";
-/// ``
-pub const DEV_BLACKBERRY: &str = "";
+/// ``
+pub const DEV_BLAZOR: &str = "";
+/// ``
+pub const DEV_BLENDER: &str = "";
 /// ``
 pub const DEV_BOOTSTRAP: &str = "";
 /// ``
 pub const DEV_BOWER: &str = "";
-/// ``
-pub const DEV_BRACKETS: &str = "";
-/// ``
-pub const DEV_BUGSENSE: &str = "";
 /// ``
-pub const DEV_CELLULOID: &str = "";
-/// ``
-pub const DEV_CHART: &str = "";
+pub const DEV_BROWSERSTACK: &str = "";
+/// ``
+pub const DEV_BULMA: &str = "";
+/// ``
+pub const DEV_BUN: &str = "";
+/// ``
+pub const DEV_C_LANG: &str = "";
+/// ``
+pub const DEV_CAIRO: &str = "";
+/// ``
+pub const DEV_CAKEPHP: &str = "";
+/// ``
+pub const DEV_CANVA: &str = "";
+/// ``
+pub const DEV_CAPACITOR: &str = "";
+/// ``
+pub const DEV_CARBON: &str = "";
+/// ``
+pub const DEV_CASSANDRA: &str = "";
+/// ``
+pub const DEV_CENTOS: &str = "";
+/// ``
+pub const DEV_CEYLON: &str = "";
 /// ``
 pub const DEV_CHROME: &str = "";
-/// ``
-pub const DEV_CISCO: &str = "";
+/// ``
+pub const DEV_CIRCLECI: &str = "";
+/// ``
+pub const DEV_CLARITY: &str = "";
+/// ``
+pub const DEV_CLION: &str = "";
 /// ``
 pub const DEV_CLOJURE: &str = "";
 /// ``
 pub const DEV_CLOJURE_ALT: &str = "";
-/// ``
-pub const DEV_CLOUD9: &str = "";
+/// ``
+pub const DEV_CLOJURESCRIPT: &str = "";
+/// ``
+pub const DEV_CLOUDFLARE: &str = "";
 /// ``
-pub const DEV_CODA: &str = "";
+pub const DEV_CLOUDFLAREWORKERS: &str = "";
+/// ``
+pub const DEV_CMAKE: &str = "";
 /// ``
-pub const DEV_CODE: &str = "";
-/// ``
-pub const DEV_CODE_BADGE: &str = "";
+pub const DEV_CODEAC: &str = "";
+/// ``
+pub const DEV_CODECOV: &str = "";
 /// ``
 pub const DEV_CODEIGNITER: &str = "";
 /// ``
 pub const DEV_CODEPEN: &str = "";
-/// ``
-pub const DEV_CODROPS: &str = "";
 /// ``
 pub const DEV_COFFEESCRIPT: &str = "";
-/// ``
-pub const DEV_COMPASS: &str = "";
 /// ``
 pub const DEV_COMPOSER: &str = "";
-/// ``
-pub const DEV_CREATIVECOMMONS: &str = "";
-/// ``
-pub const DEV_CREATIVECOMMONS_BADGE: &str = "";
+/// ``
+pub const DEV_CONFLUENCE: &str = "";
+/// ``
+pub const DEV_CONSUL: &str = "";
+/// ``
+pub const DEV_CONTAO: &str = "";
+/// ``
+pub const DEV_COREJS: &str = "";
+/// ``
+pub const DEV_COSMOSDB: &str = "";
+/// ``
+pub const DEV_COUCHBASE: &str = "";
+/// ``
+pub const DEV_COUCHDB: &str = "";
+/// ``
+pub const DEV_CPLUSPLUS: &str = "";
+/// ``
+pub const DEV_CRYSTAL: &str = "";
+/// ``
+pub const DEV_CSHARP: &str = "";
 /// ``
 pub const DEV_CSS3: &str = "";
 /// ``
 pub const DEV_CSS3_FULL: &str = "";
-/// ``
-pub const DEV_CSS_TRICKS: &str = "";
-/// ``
-pub const DEV_CSSDECK: &str = "";
+/// ``
+pub const DEV_CUCUMBER: &str = "";
+/// ``
+pub const DEV_CYPRESSIO: &str = "";
+/// ``
+pub const DEV_D3JS: &str = "";
 /// ``
 pub const DEV_DART: &str = "";
 /// ``
 pub const DEV_DATABASE: &str = "";
+/// ``
+pub const DEV_DATAGRIP: &str = "";
+/// ``
+pub const DEV_DATASPELL: &str = "";
+/// ``
+pub const DEV_DBEAVER: &str = "";
 /// ``
 pub const DEV_DEBIAN: &str = "";
+/// ``
+pub const DEV_DENOJS: &str = "";
+/// ``
+pub const DEV_DEVICON: &str = "";
 /// ``
-pub const DEV_DIGITAL_OCEAN: &str = "";
+pub const DEV_DIGITALOCEAN: &str = "";
+/// ``
+pub const DEV_DISCORDJS: &str = "";
 /// ``
 pub const DEV_DJANGO: &str = "";
+/// ``
+pub const DEV_DJANGOREST: &str = "";
 /// ``
 pub const DEV_DLANG: &str = "";
 /// ``
 pub const DEV_DOCKER: &str = "";
 /// ``
 pub const DEV_DOCTRINE: &str = "";
-/// ``
-pub const DEV_DOJO: &str = "";
 /// ``
 pub const DEV_DOTNET: &str = "";
+/// ``
+pub const DEV_DOTNETCORE: &str = "";
 /// ``
 pub const DEV_DREAMWEAVER: &str = "";
 /// ``
 pub const DEV_DROPBOX: &str = "";
+/// ``
+pub const DEV_DROPWIZARD: &str = "";
 /// ``
 pub const DEV_DRUPAL: &str = "";
+/// ``
+pub const DEV_DYNAMODB: &str = "";
 /// ``
 pub const DEV_ECLIPSE: &str = "";
+/// ``
+pub const DEV_ECTO: &str = "";
+/// ``
+pub const DEV_ELASTICSEARCH: &str = "";
+/// ``
+pub const DEV_ELECTRON: &str = "";
+/// ``
+pub const DEV_ELEVENTY: &str = "";
+/// ``
+pub const DEV_ELIXIR: &str = "";
+/// ``
+pub const DEV_ELM: &str = "";
+/// ``
+pub const DEV_EMACS: &str = "";
+/// ``
+pub const DEV_EMBEDDEDC: &str = "";
 /// ``
 pub const DEV_EMBER: &str = "";
-/// ``
-pub const DEV_ENVATO: &str = "";
+/// ``
+pub const DEV_ENVOY: &str = "";
 /// ``
 pub const DEV_ERLANG: &str = "";
-/// ``
-pub const DEV_EXTJS: &str = "";
+/// ``
+pub const DEV_ESLINT: &str = "";
+/// ``
+pub const DEV_EXPRESS: &str = "";
+/// ``
+pub const DEV_FACEBOOK: &str = "";
+/// ``
+pub const DEV_FASTAPI: &str = "";
+/// ``
+pub const DEV_FASTIFY: &str = "";
+/// ``
+pub const DEV_FAUNADB: &str = "";
+/// ``
+pub const DEV_FEATHERSJS: &str = "";
+/// ``
+pub const DEV_FEDORA: &str = "";
+/// ``
+pub const DEV_FIGMA: &str = "";
+/// ``
+pub const DEV_FILEZILLA: &str = "";
 /// ``
 pub const DEV_FIREBASE: &str = "";
 /// ``
 pub const DEV_FIREFOX: &str = "";
+/// ``
+pub const DEV_FLASK: &str = "";
+/// ``
+pub const DEV_FLUTTER: &str = "";
+/// ``
+pub const DEV_FORTRAN: &str = "";
+/// ``
+pub const DEV_FOUNDATION: &str = "";
+/// ``
+pub const DEV_FRAMERMOTION: &str = "";
+/// ``
+pub const DEV_FRAMEWORK7: &str = "";
 /// ``
 pub const DEV_FSHARP: &str = "";
+/// ``
+pub const DEV_GATLING: &str = "";
+/// ``
+pub const DEV_GATSBY: &str = "";
+/// ``
+pub const DEV_GAZEBO: &str = "";
+/// ``
+pub const DEV_GCC: &str = "";
+/// ``
+pub const DEV_GENTOO: &str = "";
 /// ``
 pub const DEV_GHOST: &str = "";
 /// ``
 pub const DEV_GHOST_SMALL: &str = "";
+/// ``
+pub const DEV_GIMP: &str = "";
 /// ``
 pub const DEV_GIT: &str = "";
 /// ``
@@ -138,245 +308,685 @@ pub const DEV_GIT_COMPARE: &str = "";
 pub const DEV_GIT_MERGE: &str = "";
 /// ``
 pub const DEV_GIT_PULL_REQUEST: &str = "";
-/// ``
-pub const DEV_GITHUB: &str = "";
-/// ``
-pub const DEV_GITHUB_ALT: &str = "";
+/// ``
+pub const DEV_GITBOOK: &str = "";
 /// ``
-pub const DEV_GITHUB_BADGE: &str = "";
+pub const DEV_GITHUB: &str = "";
 /// ``
 pub const DEV_GITHUB_FULL: &str = "";
+/// ``
+pub const DEV_GITHUBACTIONS: &str = "";
+/// ``
+pub const DEV_GITHUBCODESPACES: &str = "";
+/// ``
+pub const DEV_GITLAB: &str = "";
+/// ``
+pub const DEV_GITPOD: &str = "";
+/// ``
+pub const DEV_GITTER: &str = "";
 /// ``
 pub const DEV_GNU: &str = "";
 /// ``
 pub const DEV_GO: &str = "";
-/// ``
-pub const DEV_GOOGLE_CLOUD_PLATFORM: &str = "";
-/// ``
-pub const DEV_GOOGLE_DRIVE: &str = "";
+/// ``
+pub const DEV_GODOT: &str = "";
+/// ``
+pub const DEV_GOLAND: &str = "";
+/// ``
+pub const DEV_GOOGLE: &str = "";
+/// ``
+pub const DEV_GOOGLECLOUD: &str = "";
+/// ``
+pub const DEV_GRADLE: &str = "";
+/// ``
+pub const DEV_GRAFANA: &str = "";
 /// ``
 pub const DEV_GRAILS: &str = "";
+/// ``
+pub const DEV_GRAPHQL: &str = "";
 /// ``
 pub const DEV_GROOVY: &str = "";
+/// ``
+pub const DEV_GRPC: &str = "";
 /// ``
 pub const DEV_GRUNT: &str = "";
 /// ``
 pub const DEV_GULP: &str = "";
-/// ``
-pub const DEV_HACKERNEWS: &str = "";
+/// ``
+pub const DEV_HADOOP: &str = "";
+/// ``
+pub const DEV_HANDLEBARS: &str = "";
+/// ``
+pub const DEV_HARDHAT: &str = "";
+/// ``
+pub const DEV_HARVESTER: &str = "";
 /// ``
 pub const DEV_HASKELL: &str = "";
+/// ``
+pub const DEV_HAXE: &str = "";
+/// ``
+pub const DEV_HELM: &str = "";
 /// ``
 pub const DEV_HEROKU: &str = "";
+/// ``
+pub const DEV_HIBERNATE: &str = "";
+/// ``
+pub const DEV_HOMEBREW: &str = "";
 /// ``
 pub const DEV_HTML5: &str = "";
-/// ``
-pub const DEV_HTML5_3D_EFFECTS: &str = "";
-/// ``
-pub const DEV_HTML5_CONNECTIVITY: &str = "";
-/// ``
-pub const DEV_HTML5_DEVICE_ACCESS: &str = "";
-/// ``
-pub const DEV_HTML5_MULTIMEDIA: &str = "";
+/// ``
+pub const DEV_HUGO: &str = "";
 /// ``
 pub const DEV_IE: &str = "";
+/// ``
+pub const DEV_IFTTT: &str = "";
 /// ``
 pub const DEV_ILLUSTRATOR: &str = "";
+/// ``
+pub const DEV_INFLUXDB: &str = "";
+/// ``
+pub const DEV_INKSCAPE: &str = "";
+/// ``
+pub const DEV_INSOMNIA: &str = "";
 /// ``
 pub const DEV_INTELLIJ: &str = "";
 /// ``
 pub const DEV_IONIC: &str = "";
+/// ``
+pub const DEV_JAEGERTRACING: &str = "";
+/// ``
+pub const DEV_JAMSTACK: &str = "";
+/// ``
+pub const DEV_JASMINE: &str = "";
 /// ``
 pub const DEV_JAVA: &str = "";
-/// ``
-pub const DEV_JAVASCRIPT: &str = "";
 /// ``
-pub const DEV_JAVASCRIPT_BADGE: &str = "";
-/// ``
-pub const DEV_JAVASCRIPT_SHIELD: &str = "";
+pub const DEV_JAVASCRIPT: &str = "";
+/// ``
+pub const DEV_JAVASCRIPT_ALT: &str = "";
+/// ``
+pub const DEV_JEET: &str = "";
 /// ``
-pub const DEV_JEKYLL_SMALL: &str = "";
+pub const DEV_JEKYLL: &str = "";
 /// ``
 pub const DEV_JENKINS: &str = "";
+/// ``
+pub const DEV_JEST: &str = "";
+/// ``
+pub const DEV_JETBRAINS: &str = "";
+/// ``
+pub const DEV_JETPACKCOMPOSE: &str = "";
 /// ``
 pub const DEV_JIRA: &str = "";
-/// ``
-pub const DEV_JOOMLA: &str = "";
+/// ``
+pub const DEV_JIRAALIGN: &str = "";
 /// ``
 pub const DEV_JQUERY: &str = "";
-/// ``
-pub const DEV_JQUERY_UI: &str = "";
-/// ``
-pub const DEV_KOMODO: &str = "";
-/// ``
-pub const DEV_KRAKENJS: &str = "";
+/// ``
+pub const DEV_JSON: &str = "";
+/// ``
+pub const DEV_JULE: &str = "";
+/// ``
+pub const DEV_JULIA: &str = "";
+/// ``
+pub const DEV_JUNIT: &str = "";
+/// ``
+pub const DEV_JUPYTER: &str = "";
+/// ``
+pub const DEV_K3OS: &str = "";
+/// ``
+pub const DEV_K3S: &str = "";
+/// ``
+pub const DEV_K6: &str = "";
+/// ``
+pub const DEV_KAGGLE: &str = "";
+/// ``
+pub const DEV_KARATELABS: &str = "";
+/// ``
+pub const DEV_KARMA: &str = "";
+/// ``
+pub const DEV_KDENEON: &str = "";
+/// ``
+pub const DEV_KERAS: &str = "";
+/// ``
+pub const DEV_KIBANA: &str = "";
+/// ``
+pub const DEV_KNEXJS: &str = "";
+/// ``
+pub const DEV_KNOCKOUT: &str = "";
+/// ``
+pub const DEV_KOTLIN: &str = "";
 /// ``
-pub const DEV_KRAKENJS_BADGE: &str = "";
+pub const DEV_KRAKENJS: &str = "";
+/// ``
+pub const DEV_KTOR: &str = "";
+/// ``
+pub const DEV_KUBERNETES: &str = "";
+/// ``
+pub const DEV_LABVIEW: &str = "";
 /// ``
 pub const DEV_LARAVEL: &str = "";
+/// ``
+pub const DEV_LATEX: &str = "";
 /// ``
 pub const DEV_LESS: &str = "";
+/// ``
+pub const DEV_LINKEDIN: &str = "";
 /// ``
 pub const DEV_LINUX: &str = "";
+/// ``
+pub const DEV_LIQUIBASE: &str = "";
+/// ``
+pub const DEV_LIVEWIRE: &str = "";
+/// ``
+pub const DEV_LLVM: &str = "";
+/// ``
+pub const DEV_LODASH: &str = "";
+/// ``
+pub const DEV_LOGSTASH: &str = "";
+/// ``
+pub const DEV_LUA: &str = "";
+/// ``
+pub const DEV_LUMEN: &str = "";
 /// ``
 pub const DEV_MAGENTO: &str = "";
-/// ``
-pub const DEV_MAILCHIMP: &str = "";
+/// ``
+pub const DEV_MARIADB: &str = "";
 /// ``
 pub const DEV_MARKDOWN: &str = "";
 /// ``
 pub const DEV_MATERIALIZECSS: &str = "";
+/// ``
+pub const DEV_MATERIALUI: &str = "";
+/// ``
+pub const DEV_MATLAB: &str = "";
+/// ``
+pub const DEV_MATPLOTLIB: &str = "";
+/// ``
+pub const DEV_MAVEN: &str = "";
+/// ``
+pub const DEV_MAYA: &str = "";
 /// ``
 pub const DEV_METEOR: &str = "";
 /// ``
 pub const DEV_METEORFULL: &str = "";
-/// ``
-pub const DEV_MITLICENCE: &str = "";
-/// ``
-pub const DEV_MODERNIZR: &str = "";
+/// ``
+pub const DEV_MICROSOFTSQLSERVER: &str = "";
+/// ``
+pub const DEV_MINITAB: &str = "";
+/// ``
+pub const DEV_MITHRIL: &str = "";
+/// ``
+pub const DEV_MOBX: &str = "";
+/// ``
+pub const DEV_MOCHA: &str = "";
+/// ``
+pub const DEV_MODX: &str = "";
+/// ``
+pub const DEV_MOLECULER: &str = "";
 /// ``
 pub const DEV_MONGODB: &str = "";
-/// ``
-pub const DEV_MOOTOOLS: &str = "";
+/// ``
+pub const DEV_MONGOOSE: &str = "";
+/// ``
+pub const DEV_MOODLE: &str = "";
 /// ``
 pub const DEV_MOOTOOLS_BADGE: &str = "";
 /// ``
 pub const DEV_MOZILLA: &str = "";
-/// ``
-pub const DEV_MSQL_SERVER: &str = "";
+/// ``
+pub const DEV_MSDOS: &str = "";
 /// ``
 pub const DEV_MYSQL: &str = "";
-/// ``
-pub const DEV_NANCY: &str = "";
-/// ``
-pub const DEV_NETBEANS: &str = "";
-/// ``
-pub const DEV_NETMAGAZINE: &str = "";
+/// ``
+pub const DEV_NANO: &str = "";
+/// ``
+pub const DEV_NEO4J: &str = "";
+/// ``
+pub const DEV_NEOVIM: &str = "";
+/// ``
+pub const DEV_NESTJS: &str = "";
+/// ``
+pub const DEV_NETLIFY: &str = "";
+/// ``
+pub const DEV_NETWORKX: &str = "";
+/// ``
+pub const DEV_NEXTJS: &str = "";
 /// ``
 pub const DEV_NGINX: &str = "";
+/// ``
+pub const DEV_NGRX: &str = "";
+/// ``
+pub const DEV_NHIBERNATE: &str = "";
+/// ``
+pub const DEV_NIM: &str = "";
+/// ``
+pub const DEV_NIMBLE: &str = "";
+/// ``
+pub const DEV_NIXOS: &str = "";
 /// ``
 pub const DEV_NODEJS: &str = "";
 /// ``
 pub const DEV_NODEJS_SMALL: &str = "";
+/// ``
+pub const DEV_NODEMON: &str = "";
+/// ``
+pub const DEV_NODEWEBKIT: &str = "";
+/// ``
+pub const DEV_NOMAD: &str = "";
+/// ``
+pub const DEV_NORG: &str = "";
+/// ``
+pub const DEV_NOTION: &str = "";
 /// ``
 pub const DEV_NPM: &str = "";
-/// ``
-pub const DEV_ONEDRIVE: &str = "";
-/// ``
-pub const DEV_OPENSHIFT: &str = "";
-/// ``
-pub const DEV_OPENSOURCE: &str = "";
+/// ``
+pub const DEV_NUGET: &str = "";
+/// ``
+pub const DEV_NUMPY: &str = "";
+/// ``
+pub const DEV_NUXTJS: &str = "";
+/// ``
+pub const DEV_OAUTH: &str = "";
+/// ``
+pub const DEV_OBJECTIVEC: &str = "";
+/// ``
+pub const DEV_OCAML: &str = "";
+/// ``
+pub const DEV_OHMYZSH: &str = "";
+/// ``
+pub const DEV_OKTA: &str = "";
+/// ``
+pub const DEV_OPENAL: &str = "";
+/// ``
+pub const DEV_OPENAPI: &str = "";
+/// ``
+pub const DEV_OPENCL: &str = "";
+/// ``
+pub const DEV_OPENCV: &str = "";
+/// ``
+pub const DEV_OPENGL: &str = "";
+/// ``
+pub const DEV_OPENSTACK: &str = "";
+/// ``
+pub const DEV_OPENSUSE: &str = "";
+/// ``
+pub const DEV_OPENTELEMETRY: &str = "";
 /// ``
 pub const DEV_OPERA: &str = "";
+/// ``
+pub const DEV_ORACLE: &str = "";
+/// ``
+pub const DEV_ORY: &str = "";
+/// ``
+pub const DEV_P5JS: &str = "";
+/// ``
+pub const DEV_PACKER: &str = "";
+/// ``
+pub const DEV_PANDAS: &str = "";
 /// ``
 pub const DEV_PERL: &str = "";
-/// ``
-pub const DEV_PHONEGAP: &str = "";
+/// ``
+pub const DEV_PFSENSE: &str = "";
+/// ``
+pub const DEV_PHALCON: &str = "";
+/// ``
+pub const DEV_PHOENIX: &str = "";
+/// ``
+pub const DEV_PHOTONENGINE: &str = "";
 /// ``
 pub const DEV_PHOTOSHOP: &str = "";
 /// ``
 pub const DEV_PHP: &str = "";
+/// ``
+pub const DEV_PHPSTORM: &str = "";
+/// ``
+pub const DEV_PLAYWRIGHT: &str = "";
+/// ``
+pub const DEV_PLOTLY: &str = "";
+/// ``
+pub const DEV_PNPM: &str = "";
+/// ``
+pub const DEV_PODMAN: &str = "";
+/// ``
+pub const DEV_POETRY: &str = "";
+/// ``
+pub const DEV_POLYGON: &str = "";
+/// ``
+pub const DEV_PORTAINER: &str = "";
+/// ``
+pub const DEV_POSTCSS: &str = "";
 /// ``
 pub const DEV_POSTGRESQL: &str = "";
+/// ``
+pub const DEV_POSTMAN: &str = "";
+/// ``
+pub const DEV_POWERSHELL: &str = "";
+/// ``
+pub const DEV_PREMIEREPRO: &str = "";
+/// ``
+pub const DEV_PRISMA: &str = "";
+/// ``
+pub const DEV_PROCESSING: &str = "";
 /// ``
 pub const DEV_PROLOG: &str = "";
+/// ``
+pub const DEV_PROMETHEUS: &str = "";
+/// ``
+pub const DEV_PROTRACTOR: &str = "";
+/// ``
+pub const DEV_PULSAR: &str = "";
+/// ``
+pub const DEV_PULUMI: &str = "";
+/// ``
+pub const DEV_PUPPETEER: &str = "";
+/// ``
+pub const DEV_PURESCRIPT: &str = "";
+/// ``
+pub const DEV_PUTTY: &str = "";
+/// ``
+pub const DEV_PYCHARM: &str = "";
+/// ``
+pub const DEV_PYPI: &str = "";
+/// ``
+pub const DEV_PYSCRIPT: &str = "";
+/// ``
+pub const DEV_PYTEST: &str = "";
 /// ``
 pub const DEV_PYTHON: &str = "";
-/// ``
-pub const DEV_RACKSPACE: &str = "";
-/// ``
-pub const DEV_RAPHAEL: &str = "";
+/// ``
+pub const DEV_PYTORCH: &str = "";
+/// ``
+pub const DEV_QODANA: &str = "";
+/// ``
+pub const DEV_QT: &str = "";
+/// ``
+pub const DEV_QUARKUS: &str = "";
+/// ``
+pub const DEV_QUASAR: &str = "";
+/// ``
+pub const DEV_QWIK: &str = "";
+/// ``
+pub const DEV_R: &str = "";
+/// ``
+pub const DEV_RABBITMQ: &str = "";
+/// ``
+pub const DEV_RAILS: &str = "";
+/// ``
+pub const DEV_RAILWAY: &str = "";
+/// ``
+pub const DEV_RANCHER: &str = "";
 /// ``
-pub const DEV_RASBERRY_PI: &str = "";
+pub const DEV_RASPBERRY_PI: &str = "";
+/// ``
+pub const DEV_REACH: &str = "";
 /// ``
 pub const DEV_REACT: &str = "";
+/// ``
+pub const DEV_REACTBOOTSTRAP: &str = "";
+/// ``
+pub const DEV_REACTNAVIGATION: &str = "";
+/// ``
+pub const DEV_REACTROUTER: &str = "";
+/// ``
+pub const DEV_READTHEDOCS: &str = "";
+/// ``
+pub const DEV_REALM: &str = "";
+/// ``
+pub const DEV_RECT: &str = "";
 /// ``
 pub const DEV_REDHAT: &str = "";
 /// ``
 pub const DEV_REDIS: &str = "";
+/// ``
+pub const DEV_REDUX: &str = "";
+/// ``
+pub const DEV_RENPY: &str = "";
+/// ``
+pub const DEV_REPLIT: &str = "";
 /// ``
 pub const DEV_REQUIREJS: &str = "";
-/// ``
-pub const DEV_RESPONSIVE: &str = "";
+/// ``
+pub const DEV_RIDER: &str = "";
+/// ``
+pub const DEV_ROCKSDB: &str = "";
+/// ``
+pub const DEV_ROCKYLINUX: &str = "";
+/// ``
+pub const DEV_ROLLUP: &str = "";
+/// ``
+pub const DEV_ROS: &str = "";
+/// ``
+pub const DEV_RSPEC: &str = "";
+/// ``
+pub const DEV_RSTUDIO: &str = "";
 /// ``
 pub const DEV_RUBY: &str = "";
-/// ``
-pub const DEV_RUBY_ON_RAILS: &str = "";
 /// ``
 pub const DEV_RUBY_ROUGH: &str = "";
+/// ``
+pub const DEV_RUBYMINE: &str = "";
 /// ``
 pub const DEV_RUST: &str = "";
+/// ``
+pub const DEV_RXJS: &str = "";
 /// ``
 pub const DEV_SAFARI: &str = "";
+/// ``
+pub const DEV_SALESFORCE: &str = "";
+/// ``
+pub const DEV_SANITY: &str = "";
 /// ``
 pub const DEV_SASS: &str = "";
 /// ``
 pub const DEV_SCALA: &str = "";
-/// ``
-pub const DEV_SCRIPTCS: &str = "";
-/// ``
-pub const DEV_SCRUM: &str = "";
-/// ``
-pub const DEV_SENCHATOUCH: &str = "";
-/// ``
-pub const DEV_SIZZLEJS: &str = "";
+/// ``
+pub const DEV_SCALINGO: &str = "";
+/// ``
+pub const DEV_SCIKITLEARN: &str = "";
+/// ``
+pub const DEV_SDL: &str = "";
+/// ``
+pub const DEV_SELENIUM: &str = "";
+/// ``
+pub const DEV_SEMA: &str = "";
+/// ``
+pub const DEV_SENTRY: &str = "";
+/// ``
+pub const DEV_SEQUELIZE: &str = "";
+/// ``
+pub const DEV_SHOPWARE: &str = "";
+/// ``
+pub const DEV_SHOTGRID: &str = "";
+/// ``
+pub const DEV_SKETCH: &str = "";
+/// ``
+pub const DEV_SLACK: &str = "";
 /// ``
 pub const DEV_SMASHING_MAGAZINE: &str = "";
-/// ``
-pub const DEV_SNAP_SVG: &str = "";
+/// ``
+pub const DEV_SOCKETIO: &str = "";
+/// ``
+pub const DEV_SOLIDITY: &str = "";
+/// ``
+pub const DEV_SOLIDJS: &str = "";
+/// ``
+pub const DEV_SONARQUBE: &str = "";
+/// ``
+pub const DEV_SOURCETREE: &str = "";
+/// ``
+pub const DEV_SPACK: &str = "";
+/// ``
+pub const DEV_SPLUNK: &str = "";
+/// ``
+pub const DEV_SPRING: &str = "";
+/// ``
+pub const DEV_SPSS: &str = "";
+/// ``
+pub const DEV_SPYDER: &str = "";
+/// ``
+pub const DEV_SQLALCHEMY: &str = "";
+/// ``
+pub const DEV_SQLDEVELOPER: &str = "";
 /// ``
-pub const DEV_SQLLITE: &str = "";
+pub const DEV_SQLITE: &str = "";
+/// ``
+pub const DEV_SSH: &str = "";
 /// ``
 pub const DEV_STACKOVERFLOW: &str = "";
-/// ``
-pub const DEV_STREAMLINE: &str = "";
+/// ``
+pub const DEV_STATA: &str = "";
+/// ``
+pub const DEV_STORYBOOK: &str = "";
+/// ``
+pub const DEV_STREAMLIT: &str = "";
 /// ``
 pub const DEV_STYLUS: &str = "";
 /// ``
 pub const DEV_SUBLIME: &str = "";
+/// ``
+pub const DEV_SUBVERSION: &str = "";
+/// ``
+pub const DEV_SUPABASE: &str = "";
+/// ``
+pub const DEV_SVELTE: &str = "";
+/// ``
+pub const DEV_SWAGGER: &str = "";
 /// ``
 pub const DEV_SWIFT: &str = "";
-/// ``
-pub const DEV_SYMFONY: &str = "";
+/// ``
+pub const DEV_SWIPER: &str = "";
 /// ``
-pub const DEV_SYMFONY_BADGE: &str = "";
-/// ``
-pub const DEV_TECHCRUNCH: &str = "";
+pub const DEV_SYMFONY: &str = "";
+/// ``
+pub const DEV_TAILWINDCSS: &str = "";
+/// ``
+pub const DEV_TAURI: &str = "";
+/// ``
+pub const DEV_TENSORFLOW: &str = "";
 /// ``
 pub const DEV_TERMINAL: &str = "";
-/// ``
-pub const DEV_TERMINAL_BADGE: &str = "";
+/// ``
+pub const DEV_TERRAFORM: &str = "";
+/// ``
+pub const DEV_TEX: &str = "";
+/// ``
+pub const DEV_THEALGORITHMS: &str = "";
+/// ``
+pub const DEV_THREEDSMAX: &str = "";
+/// ``
+pub const DEV_THREEJS: &str = "";
+/// ``
+pub const DEV_TITANIUMSDK: &str = "";
+/// ``
+pub const DEV_TOMCAT: &str = "";
+/// ``
+pub const DEV_TORTOISEGIT: &str = "";
+/// ``
+pub const DEV_TOWERGIT: &str = "";
+/// ``
+pub const DEV_TRAEFIKMESH: &str = "";
+/// ``
+pub const DEV_TRAEFIKPROXY: &str = "";
 /// ``
 pub const DEV_TRAVIS: &str = "";
 /// ``
 pub const DEV_TRELLO: &str = "";
+/// ``
+pub const DEV_TRPC: &str = "";
+/// ``
+pub const DEV_TWITTER: &str = "";
+/// ``
+pub const DEV_TYPESCRIPT: &str = "";
 /// ``
 pub const DEV_TYPO3: &str = "";
 /// ``
 pub const DEV_UBUNTU: &str = "";
-/// ``
-pub const DEV_UIKIT: &str = "";
+/// ``
+pub const DEV_UNIFIEDMODELINGLANGUAGE: &str = "";
 /// ``
-pub const DEV_UNITY_SMALL: &str = "";
+pub const DEV_UNITY: &str = "";
+/// ``
+pub const DEV_UNIX: &str = "";
+/// ``
+pub const DEV_UNREALENGINE: &str = "";
+/// ``
+pub const DEV_UWSGI: &str = "";
+/// ``
+pub const DEV_V8: &str = "";
+/// ``
+pub const DEV_VAGRANT: &str = "";
+/// ``
+pub const DEV_VALA: &str = "";
+/// ``
+pub const DEV_VAULT: &str = "";
+/// ``
+pub const DEV_VERCEL: &str = "";
+/// ``
+pub const DEV_VERTX: &str = "";
 /// ``
 pub const DEV_VIM: &str = "";
+/// ``
+pub const DEV_VISUALBASIC: &str = "";
 /// ``
 pub const DEV_VISUALSTUDIO: &str = "";
-/// ``
-pub const DEV_W3C: &str = "";
-/// ``
-pub const DEV_WEBPLATFORM: &str = "";
+/// ``
+pub const DEV_VITE: &str = "";
+/// ``
+pub const DEV_VITEJS: &str = "";
+/// ``
+pub const DEV_VITESS: &str = "";
+/// ``
+pub const DEV_VITEST: &str = "";
+/// ``
+pub const DEV_VSCODE: &str = "";
+/// ``
+pub const DEV_VSPHERE: &str = "";
+/// ``
+pub const DEV_VUEJS: &str = "";
+/// ``
+pub const DEV_VUESTOREFRONT: &str = "";
+/// ``
+pub const DEV_VUETIFY: &str = "";
+/// ``
+pub const DEV_VYPER: &str = "";
+/// ``
+pub const DEV_WASM: &str = "";
+/// ``
+pub const DEV_WEBFLOW: &str = "";
+/// ``
+pub const DEV_WEBLATE: &str = "";
+/// ``
+pub const DEV_WEBPACK: &str = "";
+/// ``
+pub const DEV_WEBSTORM: &str = "";
 /// ``
 pub const DEV_WINDOWS: &str = "";
+/// ``
+pub const DEV_WINDOWS11: &str = "";
+/// ``
+pub const DEV_WOOCOMMERCE: &str = "";
 /// ``
 pub const DEV_WORDPRESS: &str = "";
-/// ``
-pub const DEV_YAHOO: &str = "";
-/// ``
-pub const DEV_YAHOO_SMALL: &str = "";
-/// ``
-pub const DEV_YEOMAN: &str = "";
+/// ``
+pub const DEV_XAMARIN: &str = "";
+/// ``
+pub const DEV_XCODE: &str = "";
+/// ``
+pub const DEV_XD: &str = "";
+/// ``
+pub const DEV_XML: &str = "";
+/// ``
+pub const DEV_YAML: &str = "";
+/// ``
+pub const DEV_YARN: &str = "";
 /// ``
 pub const DEV_YII: &str = "";
+/// ``
+pub const DEV_YUGABYTEDB: &str = "";
+/// ``
+pub const DEV_YUNOHOST: &str = "";
 /// ``
 pub const DEV_ZEND: &str = "";
+/// ``
+pub const DEV_ZIG: &str = "";