+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Unreleased
 
+## v0.3.0
+
+- Changed: Aligned with Nerd Font Symbols version `v3.3.0`.
+
 ## v0.2.0
 
-- Update: Aligned with Nerd Font Symbols version `v3.2.1`.
+- Changed: Aligned with Nerd Font Symbols version `v3.2.1`.
 
 ## v0.1.0